Delete (hopefully temporarily) repl-utils ns
After merging into oscaro's repo
(https://github.com/oscaro/ds-test-tools/pull/3)
the ci failed. specifically the repl-utils tests.

Obviously I had ran tests before PRing. The tests run fine
from CIDER inside the repl, but `lein test` from the terminal fails.
Some weird datasplash issue (`unable to deserialize Custom DoFn With Execution Info`)

There are three options :

- Fix the problem.
This would be ideal, but that I have other things to do.

- Comment out the tests.
This is not as egregiously stupid as it sounds.
The code does run as intended, and the failing is a false positive.
Furthermore, since that code is only meant for dev and never ends up
in production, it is less critical if a bug does slip in.
It does look bad though.

- Just delete the code, hopefully temporarily.
I want to use the other features (like generating all intermediate
products), so, until I have time for option 1, that's the best course
of action.